+++ duqtools/ids/dimensions.py
@@ -1,0 +1,6 @@
+def expand_dimensions():
+    pass
+
+
+def apply_operation():
+    pass
